Switch data type for transducer state from `int` to `uint8_t`

+++ src/Smarty/Transducer/AbstractTransducer.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <vector>
 
 class SmartyAbstractTransducer {
@@ -7,7 +8,7 @@ public:
   const char* getName();
   virtual bool setup() = 0;
   virtual bool loop() = 0;
-  virtual int state() = 0;
+  virtual uint8_t state() = 0;
 protected:
   const char* _name;
   SmartyAbstractTransducer(const char*);
